requirements.txt
google-api-python-client==2.77.0
pandas==1.5.3
python-dotenv==0.21.1
python-telegram-bot==13.15
Scrapy==2.8.0